designed review UI, updated fragments (not done)
