+++ .idea/misc.xml
@@ -3,7 +3,13 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
-        <entry key="app/src/main/res/layout/fragment_food_truck_detail.xml" value="0.3078125" />
+        <entry key="app/src/main/res/layout/food_item.xml" value="0.44565217391304346" />
+        <entry key="app/src/main/res/layout/food_truck_list_item.xml" value="0.4017210144927536" />
+        <entry key="app/src/main/res/layout/food_truck_review.xml" value="0.4017210144927536" />
+        <entry key="app/src/main/res/layout/fragment_food_truck_detail.xml" value="0.33" />
+        <entry key="app/src/main/res/layout/fragment_food_truck_list.xml" value="0.33" />
+        <entry key="app/src/main/res/layout/fragment_food_truck_menu.xml" value="0.4017210144927536" />
+        <entry key="app/src/main/res/layout/fragment_food_truck_reviews.xml" value="0.4017210144927536" />
       </map>
     </option>
   </component>
