+++ app/src/main/java/com/ecs198f/foodtrucks/FoodTruckDetailFragment.kt
@@ -7,6 +7,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.navArgs
 import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.viewpager2.widget.ViewPager2
 import com.bumptech.glide.Glide
 import com.ecs198f.foodtrucks.databinding.FragmentFoodTruckDetailBinding
 import retrofit2.Call
@@ -15,13 +16,14 @@ import retrofit2.Response
 
 class FoodTruckDetailFragment : Fragment() {
     private val args: FoodTruckDetailFragmentArgs by navArgs()
+    private lateinit var tabStateAdapter: TabStateAdapter
+    private lateinit var viewPager: ViewPager2
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         val binding = FragmentFoodTruckDetailBinding.inflate(inflater, container, false)
-        val recyclerViewAdapter = FoodItemListRecyclerViewAdapter(listOf())
 
         args.foodTruck.let {
             binding.apply {
@@ -29,30 +31,16 @@ class FoodTruckDetailFragment : Fragment() {
                 foodTruckDetailPriceLevel.text = "$".repeat(it.priceLevel)
                 foodTruckDetailLocation.text = it.location
                 foodTruckDetailTime.text = it.formattedTimeInterval
-                foodItemListRecyclerView.apply {
-                    adapter = recyclerViewAdapter
-                    layoutManager = LinearLayoutManager(context)
-                }
-            }
-
-            (requireActivity() as MainActivity).apply {
-                title = it.name
-
-                foodTruckService.listFoodItems(it.id).enqueue(object : Callback<List<FoodItem>> {
-                    override fun onResponse(
-                        call: Call<List<FoodItem>>,
-                        response: Response<List<FoodItem>>
-                    ) {
-                        recyclerViewAdapter.updateItems(response.body()!!)
-                    }
-
-                    override fun onFailure(call: Call<List<FoodItem>>, t: Throwable) {
-                        throw t
-                    }
-                })
             }
         }
 
         return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        // send argument to subfragment
+        tabStateAdapter = TabStateAdapter(this)
+        viewPager = view.findViewById(R.id.pager)
+        viewPager.adapter = tabStateAdapter
     }
 }
